backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.46
asyncmy==0.2.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
redis==4.1.0
python-dateutil==2.8.2
pydantic==1.8.2
aiofiles==22.1.0 
email-validator==1.3.1
cryptography==41.0.7